app/requirements.txt
# FastAPI Backend Requirements
# ArogyaKrishi - Plant Disease Detection System

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.1.0
aiosqlite==0.20.0

# Image Processing
Pillow==11.0.0
numpy==2.2.0

# Machine Learning
torch==2.7.1
torchvision==0.22.1

# Optional: LLM Integration
openai==1.59.6
anthropic==0.43.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
